requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
torch
torchvision
tensorflow
scikit-image
Pillow
pandas
